// demo/src/generated/assets.ts
export type FigmaPropertyType = 'BOOLEAN' | 'SLOT' | 'VARIANT'

export interface AssetPropertyDefinition {
  readonly figmaName: string
  readonly reactProp: string
  readonly type: FigmaPropertyType
  readonly defaultValue: string | boolean
  readonly variantOptions: readonly string[]
}

export interface AssetDefinition {
  readonly name: string
  readonly nodeId: string
  readonly type: 'COMPONENT' | 'COMPONENT_SET'
  readonly pageId: string
  readonly pageName: string
  readonly assetKey: string
  readonly componentName: string
  readonly properties: readonly AssetPropertyDefinition[]
  readonly isIcon: boolean
  readonly isInternal: boolean
}

export const assetCatalog = [
  {
    "name": "Icon/A",
    "nodeId": "64:2291",
    "type": "COMPONENT",
    "pageId": "1:2",
    "pageName": "图标 lcon",
    "assetKey": "348689bf07673f537c16f99264b5afafc82c12ae",
    "componentName": "IconA",
    "properties": [],
    "isIcon": true,
    "isInternal": false
  },
// ...
] as const satisfies readonly AssetDefinition[]